wrap in project cli style
