+++ pyproject.toml
@@ -55,6 +55,7 @@ marimapper = "marimapper.scripts.scanner_cli:main"
 marimapper_check_camera = "marimapper.scripts.check_camera_cli:main"
 marimapper_check_backend ="marimapper.scripts.check_backend_cli:main"
 marimapper_upload_mapping_to_pixelblaze = "marimapper.scripts.upload_map_to_pixelblaze_cli:main"
+marimapper_results = "marimapper.scripts.results_cli:main"
 
 
 [tool.coverage.run]
